requirements.txt
Flask==2.3.3
groq==0.4.2
PyPDF2==3.0.1
nltk==3.8.1
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0